+++ src/gpio.rs
@@ -557,6 +557,31 @@ impl<const P: char, const N: u8> IoPin<Self, Self> for Pin<Output<OpenDrain>, P,
 }
 
 impl<const P: char, const N: u8> IoPin<Pin<Input<Floating>, P, N>, Self>
+    for Pin<Output<OpenDrain>, P, N>
+{
+    type Error = Infallible;
+    fn into_input_pin(self) -> Result<Pin<Input<Floating>, P, N>, Self::Error> {
+        Ok(self.into_floating_input())
+    }
+    fn into_output_pin(mut self, state: PinState) -> Result<Self, Self::Error> {
+        self.set_state(state);
+        Ok(self)
+    }
+}
+
+impl<const P: char, const N: u8> IoPin<Self, Pin<Output<OpenDrain>, P, N>>
+    for Pin<Input<Floating>, P, N>
+{
+    type Error = Infallible;
+    fn into_input_pin(self) -> Result<Self, Self::Error> {
+        Ok(self)
+    }
+    fn into_output_pin(self, state: PinState) -> Result<Pin<Output<OpenDrain>, P, N>, Self::Error> {
+        Ok(self.into_open_drain_output_in_state(state))
+    }
+}
+
+impl<const P: char, const N: u8> IoPin<Pin<Input<Floating>, P, N>, Self>
     for Pin<Output<PushPull>, P, N>
 {
     type Error = Infallible;